[fix] rename utils.ts to config.ts
