+++ frontend/src/config.ts
@@ -1,0 +1,1 @@
+export const BASE_URL = process.env.NEXT_PUBLIC_API_URL;
